www: Fixed comment popup in prefix list
Fixed the comment tooltip popup in the prefix list. Chose to add the
tooltip to the image instead of the div as I saw some strange behaviour
when having it added to the div (a line break was inserted after the
comment image when the tooltip popped up, moving the description one row
down). Removed the explicit placement as 'bottom' is set as global
default in app.js.