nso-nipap: Update Redeployer from examples
Updated the Redeployer-class of nso-nipap to match the one in the ID and
IP allocator packages developed by tail-f/Cisco. One major change is
that a new MAAPI-connection is created for each redeploy, which seems to
have adressed the issue of NSO locking when many NIPAP requests tries to
redeploy a service in the same transaction.