Support unlimited max_result for search_prefix
For certain queries, like when expanding a prefix by clicking the '+' in
the web UI we really shouldn't have a limit. Unlimited is signified by
setting max_result to False or None.

Downside might be DoS attack on the backend or some incorrect logic
somewhere blowing it up with heavy queries inadvertently.

Fixes #466.