Integrate DB install / upgrade in nipapd
The debian postinst script for nipapd has typically been responsible for
installing and/or upgrading the database schema to match the latest
version but it has certain drawbacks;
 * only works on debian
 * doesn't share connection parameters with nipapd

By integrating the db schema install/upgrade functionality into nipapd
we get a portable solution that should work on any platform that can run
nipapd and if it can't run nipapd it's probably not very interesting
installing the db schema anyway. In addition, the database connection
parameters are read from the nipap.conf configuration file. Unlike the
old postinst script, which assumed the database to be run on the local
machine, we can now connect to a remote database and install our schema
there.

There are two options;
 * --auto-install-db
 * --auto-upgrade-db

The rationale being that it might be a no-brainer to install the schema
if the database is empty (fresh install) but an upgrade potentially
means some downtime (slow upgrade process) and thus we have two options.

Just like before, downgrades are not supported at all. We usually add
columns between db schema versions and a downgrade would thus mean
deleting columns which means data loss. If someone truly does want to
downgrade, that will have to be done manually.