Rework debian install scripts
Since nipapd now takes care of database schema installation and upgrades
we no longer need to do this in the debian install scripts. The logic
has thus been vastly simplified but on the other hand I've also added
support for having a remote database which increases the complexity
somewhat. Not sure what the net effect is ;)