cli: read connection params from env or prompt
In addition to reading the connection parameters from the .nipaprc
configuration file we now also try to get them from environment
variables. We prefer environment variables, then from .nipaprc and last,
in the case of the password, we will interactively prompt for it.

nipapd host and port now also have the defaults 'localhost' and 1337
respectively.

Fixes #838.