backend: interpret VRF RTs as VRF RTs
Previously we interpreted two colon separated numbers as just a string
and would try to match in description, order_id, comment and so forth.
There was however a bug and we would try to match on 'name' which isn't
a valid column / attribute for prefixes so we'd get an error.
We now switch over to match in the vrf_rt column instead.

Fixes #952.