backend: switch nipap.conf.dist to Jinja2 syntax
I've switched out some of the values that we want to substitue to Jinja2
style variables so that it's easier to use, well, Jinja2 to do the
rendering of the config file.

This is kind of preparatory work for putting nipapd in a docker
container as I envision using envtpl to do the variable subsitution.

The debian nipapd postinst script is however still using sed since I
didn't want to include envtpl in the normal debian package and I'm too
lazy to write a Jinja2 renderer right now. This should still be better
than before IMHO.

Fixes #949.