cli: Removed unnecessary list conversions
Probably to be on the safe side, the 2to3 tool seems to wrap all uses of
.items() in a list() to convert the result to a list. This is due to an
API change, described briefly at
https://docs.python.org/3/whatsnew/3.0.html#views-and-iterators-instead-of-lists,
which might cause trouble at some occations. As we're iterating the
result from the .items()-call we should be safe without the list
conversion and they are thus removed.