backend: Fix format of smart_search result on err
Fixes the format of the return value from the smart_search-functions
when encountering smart parsing errors.

I also added some "empty" (value None) fields to simplify client
development, IMHO it's better that the keys in the response dict exist
(with a value of None) that that they sometimes do and sometimes don't.