backend: Introduce success return val from parser
Introduced a success return value from the smart parser, signalling
whether the parsing was successful or not. With this change, the code
which creates the interpretation result returned when an error occurrs
is also moved from the backend module to the smart_parsing module.

As there is no smart parser implemented for AS numbers, I've hard-coded a
successful response in the old-style query parser code.