backend: Verify enum values in smart parser
Verify that the value provided in a smart parser expression matching
against an enum value is valid.

For example, currently the only prefix types permitted are 'assignment',
'host' and 'reservation'. The changes introduced by this commit verifies
that any expression matching against the prefix attribute 'type' is one
of these values and returns an error otherwise (instead of forwarding a
database exception which was the case before).

Fixes #972