backend: Modified test case for query failure
Modified the test case looking into the entire result from a
smart_search query by adding aa query error indicator. It's function is
to provide a quick way to verify that a query has been successful (so
far only dealing with the parsing of the query string). This would
otherwise require the client to search through the entire interpretation
data structure, which potentially is pretty large.