backend: Added error attr to smart search response
Added an 'error' attribute to the response of the smart search
functions. It's function is to provide a quick way to verify whether a
query has been successful or not (so far only dealing with the parsing
of the query string). This would otherwise require the client to search
through the interpretation data structure.