www: URL encode JSON strings
Apparently the AngularJS $http-function does not do URL encoding of the
params passed to it (https://github.com/angular/angular.js/issues/8064).

As this is needed to handle exotic characters such as ";", the AngularJS
controllers has been modified to URL encode the JSON strings sent to the
XHR functions and the XHR functions has been modified to decode the data
received.

Fixes #934