www: Refactor handling of strings in request
Refactor the code handling string values in the request. Created a
function which verifies if the value provided is a non-empty string and
if so returns it strip():ed. If not, None is returned. As the code
handling this previously varied slightly between values and
add/edit-functions, the behaviour of the web UI when receiving
empty values migt have changed slightly.

Hopefully, it's for the better :)

It was noticed that the type of the values in the JSON data received by
the backend varied slightly in that a field which contained a *digit
provided by the backend* (such as a numeric VLAN when *editing* a prefix
- the value was provided by the backend) would be of type int and not
String. This only seems to happen to values stored as integers in the
database and not when for example setting a description to "7" which
would cause the string validator function to replace the value with
None. Still, feels a bit scary. It's probably due to the value (which
was received as an integer from the backend) is stored by AngularJS in
an internal data structure and not actually fetched from the DOM until
changed, when the type would be String.