www: Handle top-level error flag from smart_search
The "top-level" error flag newly introduced in the
smart_search_-functions collided with an error flag used by the web
interface. This is a quick fix for that issue which simply removes the
error key from the result received from smart_search. The flag is
currently only used to signal query string parser errors, which also are
present deeper in the interpretation dict where it will be picked up and
displayed by the web UI. Thus no functionality is lost when the
top-level flag is removed.

I was probably a bit too quick implementing the error flag; we should
probably think through how to convey "soft errors" (such as string
parsing errors) vs "hard errors" (such as, for example, the backend
being unable to reach the database) and how to do it in a future
JSON-based REST API.