www: Add dockerfile and entrypoint for nipap-www
Added a dockerfile and entrypoint-script for nipap-www. It's based on a
standard ubuntu:xenial-image, serving the NIPAP web UI with apache2 with
mod_wsgi.

All installation of dependencies and NIPAP packages is done in the
dockerfile, while the local auth database is setupi, permissions are
corrected and apache2 configured in the entrypoint shell script. The
dockerfile currently doesn't provide any way to make changes to the
nipap-www.ini-file.