nipapd: mangle IPv4 address
This mangles IPv4 address in search queries through IPy to get a
properly formatted IPv4 address before we send it to Postgres. We do
validation of IPv4 addresses but once they are verified we send the
original value, so for example 10.18.01.1 is recognized as IPv4 but
Postgres doesn't like the leading 0 in the third octet. By running it
through IPy we get rid of that leading 0 and all is fine.

Fixes #860.