www: Add slash to generated URI
After the AngularJS upgrade from v1.3.15 to v1.5.3 a slash has been
introduced in the URI from some reason. The code in NIPAP which stores
the current search state in the URI neeed to be updated to reflect this,
to avoid the page from reloading when opening the VRF selector.