smart_parsing: Unicode friendly log messages
Replace all use of str() with unicode().

Python2.7 and unicode.. sigh. We use str() in a couple of places but
this tries to cast stuff to US ASCII so anything that's not will throw
an exception. It's a tad funny how parts of this code really does handle
unicode well but it's the logging actions that break due to the use of
str().

Part of #1035.