smart_parsing: fix unicode search
We need to match on unicode characters and not just US ASCII, which the
old alphanums did. Since it's really difficult building an inclusive
list of unicode characters I took the other approach and build a regex
that matches everything except or a few "reserved" characters that we
use as operators.

Fixes #1035.