nipapd: Fix override of configuration params
With the change from optparse to argparse, the override of configuration
parameters from the config file with CLI arguments stopped working. This
was due to argparse assigning the default value "False" to arguments
with action store_true, while optparse assigned None. This was fixed by
explicitly setting the default value to None.

Also did some cleanup of the code where the override is performed.