www: Re-enable parsing of query string in URL
Since the last upgrade of AngularJS, the part of the URL to the right of
the hash has been prefixed with a slash. The library used to parse the
URL considers the slash as part of the name of the parameter and thus
the name needs to be updated when accessing it.

With this change, the URL is again succesfully parsed.

This is a hack to re-enable the functionality lost with the AngularJS
upgrade. This will all be replaced with an AngularJS implementation of
the prefix list page.