Change CI test environment to Ubuntu Bionic in LXD
Change the CI test environment to run in an LXD container running Ubuntu
Bionic. This included changing to PostgreSQL 10, as that's what's
shipped with Bionic, and removing an explicit upgrade of dpkg which was
added to work around a bug in Ubuntu Trusty.

Bump CI test env to PostgreSQL 10.

Add rename used when building .debs and remove --force-yes from apt-get
as it's considered dangerous and deprecated.