www: Improvements of VRF list smart search
Althoug the smart search functionality was implemented, the VRF list
page still performed a slow bulk-load of all VRFs. Replaces this with an
initial smart search query which matches any VRF and returns the 50
first entries to begin populating the list.

Also disabled infinite-scroll while data is fetched to avoid bursting
requests when reaching the end of the page, added a loading indicator
and stopped the infinity scroll when all VRFs has been fetched.