nipapd: Prohibit removing more pfxs than expected
Fixed a bug where a recursive remove might remove more prefixes than the
user asks for, if prefix type was used in the filter. To handle database
constraints, host prefixes are removd first in a recursive remove
operation, but the original type was never restored in the filter. Now
it is.