www: Moved stylesheets and remove workaround
To make sure bootstrap's relative paths work properly, it was moved to a
separate css-subfolder. That's not a bad idea, so all other stylesheets
were moved as well. That also meant that the workaround in place to
serve images from /images and fonts from /fonts was needed any more and
thus removed.