nipapd: Replace "" with nothing
There was a special case in the deb package postinst action which set
the psql database host to "" if it was configured as localhost.  Not
sure why this was needed, but "" caused psycopg2 to break in Python 3,
could be due to different handling in the config parser... Trying to
write nothing instead of "".