backend: Handle db_host = ""
The behavour of configparser seems to have changed, where the Python 2
version returned an empty string when a value in the config file was set
to "", two double quotes, and the Python 3 version returns a string
containing two double quotes which the PostgreSQL client library
dislikes as database host. As this is the default config generated by
the Debian install scripts it's probably pretty common, so a workaround
was implemented where the string "" results in the previous behaviour.