nipapd: Fork also when running in foreground
Since forever nipapd have avoided to fork multiple processes when
running in forground. I think the idea was to simplify debugging etc,
never to be used in production. However, when deploying nipapd as a
Docker container the process is running is foreground by default also in
production and thus the behaviour is now changed to fork also for that
case.